--- src/BoardStatus.tsx
@@ -1,10 +1,0 @@
-export const boardDefault = [
-  ["PLAYER", "COLLEGE", "YEAR", "POS", "RND", "PICK"],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-  ["", "", "", "", "", ""],
-];
