+++ src/components/Player.ts
@@ -1,4 +1,4 @@
-interface Player {
+export interface Player {
     name: string;
     college: string;
     year: number;
