+++ src/components/pop-ups/game-over/GameOver.css
